Update 2.7: 1. Quantity is persisted as a float in the database now. 2. Updated documentation and flow diagrams 3. Using pseudorandom integer generation with seedrandom 4. Updated ui-assessment with appropriate pictures 5. MatrixEntry class now has it's own ts file 6. View now communicates to the domain model indirectly via the controllers. 7. Added the new Auto-Buy feature. 8. Renamed cashier's cart property to currentCart to better reflect the relationship between that cashier and cart instance. 9. We now have a export function in view for showing errors in dialogs instead of repeating the same method in each view class.

+++ create-tables.sql
@@ -16,7 +16,7 @@ create table if not exists product(
     price integer not null,
     product_type varchar(255) not null,
 
-    quantity integer,
+    quantity float,
     
     cart_id integer not null,
     foreign key (cart_id) references cart(id) on delete cascade
